feat: cleaner YAML interface
This updates several aspects of the YAML interface, including making
likelihoods/components into lists of dicts for simpler parsing. It also
makes the data IO more explicit (must be tagged in the YAML) and
thus less prone to error.

It also adds the `Params` class to provide an interface to a group
of active parameters, and easier ability to output those to a YAML
config etc. It removes alll `child_active_params_dct` attributes,
since the new `Params` class is both like a list and a dict at the same
time.

BREAKING CHANGE: YAML API is quite different and no longer
a `child_active_params_dct` attribute in components/likelihoods.