fix: support for improper uniform priors with correct pdf
This fix adds several critical things. It ensures that if
using min/max on a parameter, the prior reflects that range.
It also implements proper priors for polychord, which before
required setting min/max but now can use arbitrary proper
priors. It also fixes a problem with priors defined on
LikelihoodContainer, which double-counted parameters shared
between likelihoods. It then makes a few fixes to the API.

BREAKING CHANGE: the API has remained very similar, but the
  output logprior will be very different in the default case,
  of a proper uniform.